+++ packages/gestalt-codemods/1.26.0-1.27.0/__testfixtures__/avatar-update-sizes-remove-icon-icon.input.js
@@ -1,0 +1,14 @@
+// @flow
+import React from 'react';
+import { Avatar } from 'gestalt';
+
+export default function Example() {
+  return (
+    <Avatar
+      src="https://i.ibb.co/ZfCZrY8/keerthi.jpg"
+      verified
+      icon="pinterest"
+      name="Keerthi"
+    />
+  );
+}